Fixed centred nav bar and tile flip misalignment
These changes have fixed the nav bar not appearing across the screen and the misalignment when flipping the information tiles over.

1) created copy pages to work on the file in question without editing main files.

2) in style.css I have changed the way items were aligned within the page
- removed alignment for whole body and header - this was affecting the nav bar as it is included in the body tags.
- edited the .group tags to center each group and used flex-wrap to wrap the second group onto the next line.
- aligning the groups this way has removed the tiles moving slightly out of line when flipping over.

nav bar and heading are still some px out of line when switching between tabs - i can't find the cause of this atm.

my next steps are to rename the page so it replaces "topics" and clean up the files and codebase.